backend/requirements.txt
# FastAPI и зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Безопасность и аутентификация
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pyotp==2.9.0
qrcode==7.4.2

# Валидация
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS и middleware
python-dotenv==1.0.0

# Логирование
python-json-logger==2.0.7

